Adds the ability to update ACEs when username changes
This commit adds an updateUserSecurityIdentity method, which is necessary when a user updates their username.

A deleteAceBySecurityIdentity was also added. This allows you to pass in a UserSecurityIdentity object and delete all ACEs associated with it. This is useful when editing a user's permissions via checkboxes, for example, and being able to delete all their ACEs and recreate them.